requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
pdfplumber>=0.10.2
pytesseract>=0.3.10
python-dotenv>=1.0.0
transformers>=4.41.1
torch>=2.3.0
Pillow>=10.3.0
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.1
